feat: fix usewindowresize hook (#292)
